requirements.txt
streamlit
python-binance
pandas
ta
